Trying to fix laGP so it won't estimate nugget

+++ R/IGP_GPfit.R
@@ -121,10 +121,14 @@ IGP_GPfit <- R6::R6Class(classname = "IGP_GPfit", inherit = IGP_base,
 #'   updates the model, adding new data if given, then running optimization again.}}
 IGP_laGP <- R6::R6Class(classname = "IGP_laGP", inherit = IGP_base,
                             public = list(
-                              .init = function(...) {
+                              .init = function(..., d=NULL, g=NULL, theta=NULL, nugget=NULL, estimate_params=TRUE) {browser()
                                 if (self$corr[[1]] != "gauss") {
                                   stop("laGP only uses Gaussian correlation")
                                 }
+
+                                if (is.null(d) & !is.null(theta)) {d <- 1/theta}
+                                if (is.null(g) & !is.null(nugget)) {g <- nugget}
+
                                 da <- laGP::darg(list(mle=TRUE), X=self$X)
                                 ga.try <- try(ga <- laGP::garg(list(mle=TRUE), y=self$Z), silent = T)
                                 if (inherits(ga.try, "try-error")) {
@@ -135,18 +139,40 @@ IGP_laGP <- R6::R6Class(classname = "IGP_laGP", inherit = IGP_base,
                                 # Follow recommendations for small samples, otherwise use bigger range
                                 drange <- if (nrow(self$X)<20) c(da$min, da$max) else c(1e-3,1e4) #c(da$min, da$max), # Don't like these small ranges
                                 grange <- c(ga$min, ga$max)
-                                mod1 <- laGP::newGPsep(X=self$X, Z=self$Z, d=da$start, g=ga$start, dK = TRUE)
+                                da_start <- if (!is.null(d)) d else da$start
+                                ga_start <- if (!is.null(g)) g else ga$start
+                                mod1 <- laGP::newGPsep(X=self$X, Z=self$Z, d=da_start, g=ga$start, dK = TRUE)
                                 #mod1 <- laGP::newGPsep(X=X, Z=Z, d=da$start, g=1e-6, dK = TRUE)
-                                mle.out <- laGP::jmleGPsep(gpsepi = mod1,
-                                                           drange=drange,
-                                                           grange=grange,
-                                                           #dab=da$ab, gab=ga$ab, # Will use MLE without these
-                                                           verb=0, maxit=1000)
-                                self$mod.extra$theta = as.numeric(1 / mle.out[1,1:ncol(self$X)]) # store theta params
-                                self$mod.extra$nugget = as.numeric(mle.out[1,ncol(self$X) + 1]) # store nugget
+                                if (!estimate_params) { #using d and g given
+                                  self$mod.extra$theta = as.numeric(1 / d) # store theta params
+                                  self$mod.extra$nugget = as.numeric(g) # store nugget
+                                } else if (estimate_params && !self$estimate.nugget) { # Only estimate d/theta
+                                  mle.out <- laGP::mleGPsep(gpsepi = mod1,
+                                                            param="d",
+                                                             tmin=drange[1], tmax=drange[2],
+                                                             verb=0, maxit=1000)
+                                  self$mod.extra$theta <- as.numeric(1 / mle.out$d) # store theta params
+                                  self$mod.extra$nugget <- nugget
+                                  # Leave nugget as is
+                                } else if (estimate_params) {
+                                  mle.out <- laGP::jmleGPsep(gpsepi = mod1,
+                                                             drange=drange,
+                                                             grange=grange,
+                                                             #dab=da$ab, gab=ga$ab, # Will use MLE without these
+                                                             verb=0, maxit=1000)
+                                  self$mod.extra$theta = as.numeric(1 / mle.out[1,1:ncol(self$X)]) # store theta params
+                                  self$mod.extra$nugget = as.numeric(mle.out[1,ncol(self$X) + 1]) # store nugget
+                                } else {stop("Shouldn't be here IGP_laGP #32097555")}
                                 self$mod <- mod1
                               }, #"function to initialize model with data
-                              .update = function(...) {
+                              .update = function(..., estimate_params=TRUE) {browser()
+                                if (!estimate_params) { # just add data and return
+                                  laGP::updateGPsep(gpsepi=self$mod,
+                                                    X=self$X[-(1:self$n.at.last.update), , drop=FALSE],
+                                                    Z=self$Z[-(1:self$n.at.last.update)])
+                                  return()
+                                }
+
                                 # Start over if not many points, had problems getting stuck in bad spots early
                                 if (self$n.at.last.update < 20) {
                                   self$.delete()
@@ -175,24 +201,44 @@ IGP_laGP <- R6::R6Class(classname = "IGP_laGP", inherit = IGP_base,
                                 }
                                 drange <- c(1e-3,1e4)
                                 grange <- c(min(sqrt(.Machine$double.eps),self$mod.extra$nugget), max(1,self$mod.extra$nugget))
-                                mle.try <- try(mle.out <- laGP::jmleGPsep(gpsepi = self$mod,
-                                                                          #drange=c(da$min, da$max), # Getting rid of these here too
-                                                                          #grange=c(ga$min, ga$max),
-                                                                          drange=drange,
-                                                                          grange=grange, # Had error of nugget starting outside bound
-                                                                          #dab=da$ab, gab=ga$ab,
-                                                                          verb=0, maxit=1000))
-                                if (inherits(mle.try, "try-error")) {
-                                  # Sometimes gives error: L-BFGS-B needs finite values of 'fn'
-                                  browser()
-                                  warning('Restarting laGP model')
-                                  self$delete()
-                                  self$init(...)
-                                  return()
+                                if (!estimate_params) {
+                                  stop("This is covered above at beginning of .update")
+                                } else if (estimate_params && !self$estimate.nugget) { # update d/theta but not nugget/g
+                                  mle.try <- try(mle.out <- laGP::mleGPsep(gpsepi = self$mod,
+                                                                           param = "d",
+                                                                            tmin=drange[1], tmax=drange[2],
+                                                                            verb=0, maxit=1000))
+                                  if (inherits(mle.try, "try-error")) {
+                                    # Sometimes gives error: L-BFGS-B needs finite values of 'fn'
+                                    browser()
+                                    warning('Restarting laGP model')
+                                    self$delete()
+                                    self$init(..., estimate_params=estimate_params)
+                                    return()
+                                  }
+                                  # Update stored parameters for when user calls $theta() or $nugget()
+                                  self$mod.extra$theta = as.numeric(1 / mle.out$d) # store theta params
+                                  # leave nugget as it was
+                                } else if (estimate_params && self$estimate.nugget) {
+                                  mle.try <- try(mle.out <- laGP::jmleGPsep(gpsepi = self$mod,
+                                                                            #drange=c(da$min, da$max), # Getting rid of these here too
+                                                                            #grange=c(ga$min, ga$max),
+                                                                            drange=drange,
+                                                                            grange=grange, # Had error of nugget starting outside bound
+                                                                            #dab=da$ab, gab=ga$ab,
+                                                                            verb=0, maxit=1000))
+                                  if (inherits(mle.try, "try-error")) {
+                                    # Sometimes gives error: L-BFGS-B needs finite values of 'fn'
+                                    browser()
+                                    warning('Restarting laGP model')
+                                    self$delete()
+                                    self$init(...)
+                                    return()
+                                  }
+                                  # Update stored parameters for when user calls $theta() or $nugget()
+                                  self$mod.extra$theta = as.numeric(1 / mle.out[1,1:ncol(self$X)]) # store theta params
+                                  self$mod.extra$nugget = as.numeric(mle.out[1,ncol(self$X) + 1]) # store nugget
                                 }
-                                # Update stored parameters for when user calls $theta() or $nugget()
-                                self$mod.extra$theta = as.numeric(1 / mle.out[1,1:ncol(self$X)]) # store theta params
-                                self$mod.extra$nugget = as.numeric(mle.out[1,ncol(self$X) + 1]) # store nugget
                               }, #"function to add data to model or reestimate params
                               .predict = function(XX, se.fit, ...){
                                 if (se.fit) {
